// src/api/Order/endpoints.ts
export enum OrderApi {
  GetAll = "Order/GetAll?Status=",
  GetConfirmAdd = "Order/GetConfirmAdd",
  GetConfirmBranches = "Order/GetConfirmBranches",
  GetConfirmDrivers = "Order/GetConfirmDrivers",
  GetConfirmReceiveCustomer = "Order/GetConfirmReceiveCustomer",
  GetConfirmReceiveBranches = "Order/GetConfirmReceiveBranches",
  GetComplete = "Order/GetComplete",
  GetInComplete = "Order/GetInComplete",
  ConfirmAdd = "Order/ConfirmAdd",
  Handle = "Order/Handle",
  ConfirmBranch = "Order/ConfirmBranch",
  ConfirmDriver = "Order/ConfirmDriver",
  ConfirmReceiveBranch = "Order/ConfirmReceiveBranch",
  ConfirmReceiveCustomer = "Order/ConfirmReceiveCustomer",
  RoleBack = "Order/RoleBack",
  Cancel = "Order/Cancel",
  UploadBill = "Order/UploadBill",
  Delete = "Order/Delete",
}