// src/enums/OrderStage.ts
export enum Order_Stage {
  NewOrder = 'NewOrder',
  Confirmed = 'Confirmed',
  UnConfirmed = 'UnConfirmed',
  ConfirmedBranches = 'ConfirmedBranches',
  ConfirmedDriver = 'ConfirmedDriver',
  ReceivedFromBranches = 'ReceivedFromBranches',
  ReceivedToCustomer = 'ReceivedToCustomer',
  Cansel = 'Cansel',
  RollBac = 'RollBack',
}